LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-table-002-expected.txt

top




bottom
top




bottom

PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="65"><span></span></div>
offsetLeft expected 65 but got 115
PASS .target > * 4
